Unit testing Config

+++ Java/src/main/java/com/carmasys/carmalinksdk/Config.java
@@ -19,8 +19,8 @@ public class Config {
 		DEACTIVATED;
 	}
 	
-	private Integer threshold;
-	private Double allowance;
+	private Float threshold;
+	private Float allowance;
 	private BuzzerVolume buzzer;	
 	private Boolean location;
 	private Status status;
